Overflow you say? Nay says I!

+++ libSpatialSEIR/include/DataStructures/CompartmentalModelMatrix.hpp
@@ -33,7 +33,7 @@ namespace SpatialSEIR
             int genFromText(std::string filename);
             int genFromDataStream(int *indata, int *inrow, int *incol);
             int createEmptyCompartment(int *inrow, int *incol);
-            int marginSum(int margin, int slice);
+            long unsigned int marginSum(int margin, int slice);
             ~CompartmentalModelMatrix();
 
             // Attributes
